requirements.txt
aiomysql==0.2.0
alembic==1.14.0
APScheduler==3.11.0
asyncpg==0.30.0
bcrypt==4.2.1
black
boto3==1.35.90
cryptography==44.0.0
Faker==35.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastapi-mail==1.4.2
greenlet==3.1.1
httptools==0.6.4
httpx==0.28.1
meilisearch==0.34.0
orjson==3.10.13
packaging==24.2
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
SQLAlchemy-Utils==0.41.2
